requirements.txt
fastapi
uvicorn
pillow
paho-mqtt
